May fixed breakpoint-problem
Problem:
After debugsession started already given breakpoints couldn't be removed
and the debugger stopped anyway on the breakpoint positions.
It seems that this was caused by the debugger loosing the connection
to the old breakpoints which could then not be removed anymore!

Fix:
Removing all breakpoints before setting the breakpoints "again"
(and so the "override" of an old breakpoint) avoids the overlayed break-
points.

- Added "del" to launchCommands in launch.json
 -> del deletes all breakpoints